ELEC-119 uint16 -> uint64 for CAN data.

+++ pkg/msgs/can_msgs.go
@@ -8,10 +8,10 @@ import (
 type CAN struct {
 	ID        uint16
 	Timestamp time.Time
-	Data      uint16
+	Data      uint64
 }
 
 // NewCAN creates a CAN struct, this should be used for all incoming messages.
-func NewCAN(id uint16, data uint16) CAN {
+func NewCAN(id uint16, data uint64) CAN {
 	return CAN{id, time.Now().UTC(), data}
 }
