+++ pkg/sources/fake/fake.go
@@ -18,7 +18,7 @@ var startTime = time.Now()
 func getCAN(id uint16) msgs.CAN {
 	t := time.Since(startTime)
 	fid := float64(id)
-	data := uint16(33000 + 3000*math.Sin((t.Seconds()/20)+fid))
+	data := uint64(33000 + 3000*math.Sin((t.Seconds()/20)+fid))
 
 	return msgs.NewCAN(id, data)
 }
